requirements.txt
anthropic==0.86.0
python-dotenv==1.2.2
requests==2.33.0
